backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
beautifulsoup4==4.14.3
bidict==0.23.1
blinker==1.9.0
bs4==0.0.2
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Deprecated==1.3.1
Flask==3.1.2
flask-cors==6.0.1
Flask-Limiter==4.0.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
limits==5.6.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy==2.2.6
oauthlib==3.3.1
opencv-python==4.12.0.88
ordered-set==4.1.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pygame==2.6.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.12.3
python-socketio==5.15.0
python-whois==0.9.3
dnspython==2.6.1
pytz==2025.2
qrcode==8.2
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.3
simple-websocket==1.1.0
six==1.17.0
soupsieve==2.8
SQLAlchemy==2.0.44
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.4
wrapt==2.0.1
wsproto==1.3.2
reportlab==4.0.9
pdfkit==1.0.0
slack-sdk==3.23.1
requests-futures==1.0.0
cryptography==42.0.2
stix2==3.0.0
APScheduler==3.10.4